% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_all_vector_combinations.R
\name{code_all_vector_combinations}
\alias{code_all_vector_combinations}
\title{All Combinations of a List}
\usage{
code_all_vector_combinations(vector, include.these.in.each = NULL)
}
\arguments{
\item{vector}{the list of items}

\item{include.these.in.each}{items that you want to make sure are included in each returned itme}
}
\description{
Returns all possible combinations of a list optionally appending some elements to make sure they are included in each list item
}
\keyword{automation}
\keyword{code}
