% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_csv_list_to_vector.R
\name{code_csv_list_to_vector}
\alias{code_csv_list_to_vector}
\title{Adapt CSV List to Vector}
\usage{
code_csv_list_to_vector(csv_list)
}
\arguments{
\item{csv_list}{comma-seperated list in character format}
}
\description{
Returns a vector
}
\examples{

}
\keyword{coder,}
\keyword{sql}
