% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_group_by_list.R
\name{code_group_by_list}
\alias{code_group_by_list}
\title{SQL Group by list}
\usage{
code_group_by_list(group_by_cols)
}
\arguments{
\item{group_by_cols}{The columns that will be included in the group by statements}
}
\description{
Turns a vector into a list of group by statements that capture every combination of items in the vectory
}
