% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_make_filter_list.R
\name{code_make_filter_list}
\alias{code_make_filter_list}
\title{Filter List Shortcut}
\usage{
code_make_filter_list(dataset)
}
\description{
Makes a filter list
}
\details{
Makes a list based on the columns in a dataset that can be used
in a WHERE statement
}
