% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_make_select_list.R
\name{code_make_select_list}
\alias{code_make_select_list}
\title{Select List Shortcut}
\usage{
code_make_select_list(dataset)
}
\description{
Makes a select list
}
\details{
Makes a list based on the columns in a dataset that can be used
in a SELECT statement
}
