% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_case_statements.R
\name{code_sql_case_statements}
\alias{code_sql_case_statements}
\title{SQL Case Statements}
\usage{
code_sql_case_statements(df)
}
\arguments{
\item{df}{dataframe with rows that will be used to create the case statements}
}
\description{
Builds case statements to use in a SELECT statement based on information in a dataframe
}
\details{
This makes it easier to seperate sample definitions
}
