% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_count.R
\name{code_sql_count}
\alias{code_sql_count}
\title{Count Records}
\usage{
code_sql_count(sql)
}
\arguments{
\item{sql}{an object of type sql_sample, character in SQL format or a db_table that has code which selects a sample of records}
}
\description{
SQL Code to get a distinct column values
}
\keyword{database,}
\keyword{location}
\keyword{state,}

