% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_distinct.R
\name{code_sql_distinct}
\alias{code_sql_distinct}
\title{Distinct Records}
\usage{
code_sql_distinct(sql, col)
}
\arguments{
\item{col}{the columns that you want to look at}

\item{db_table}{an object of type db_table which includes the table where the records are located}
}
\description{
SQL Code to get a distinct column values
}
\keyword{database,}
\keyword{location}
\keyword{state,}

