% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_head.R
\name{code_sql_head}
\alias{code_sql_head}
\title{Get Top Records}
\usage{
code_sql_head(sql, num_rows = 5)
}
\arguments{
\item{sql}{an object of type {\link{code_sql_sample}}, character in SQL format that has SQL which selects records}
}
\description{
SQL Code to get the top five records from a sample
}
\keyword{database,}
\keyword{location}
\keyword{state,}

