% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_join.R
\name{code_sql_join}
\alias{code_sql_join}
\title{SQL JOIN}
\usage{
code_sql_join(sql, table_to_join, by, sel.cols = "*", join_type = "inner")
}
\arguments{
\item{sql}{an object of type {\link{code_sql_sample}}, character in SQL format that has SQL which selects records}

\item{table_to_join}{the table that you want to join to}

\item{by}{the SQL string criteria that joins the records (ie: X.a = Y.a)}

\item{sel.cols}{the columns from the supplied sql that you would like to select}

\item{join_type}{the type of join (ie: inner, left, right)}
}
\description{
SQL code to join table
}
\details{
Joins table to sql query
}
\keyword{sql}

