% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_query.R
\name{code_sql_query}
\alias{code_sql_query}
\title{Code a SQL query}
\usage{
code_sql_query(table.name, select.cols = c("count(*) AS N"), joins = NULL,
  filters = NULL, filters.logical = "AND", group.by.cols = NULL,
  order.by.cols = NULL, alias = "RECORDS")
}
\description{
Returns a formated sql query based on parameters. Useful when automating code genaration
}
\keyword{coder,}
\keyword{query}
\keyword{sql,}

