% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_select.R
\name{code_sql_select}
\alias{code_sql_select}
\title{Code a SQL SELECT}
\usage{
code_sql_select(table.name, select.cols = c("count(*) AS N"), joins = NULL,
  filters = NULL, filters.logical = "AND", group.by.cols = NULL,
  order.by.cols = NULL, alias = "RECORDS")
}
\description{
Returns a formated sql SELECT query based on parameters. Useful when automating code
}
\keyword{coder,}
\keyword{query}
\keyword{sql,}

