% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_vector_to_csv_list.R
\name{code_vector_to_csv_list}
\alias{code_vector_to_csv_list}
\title{Adapt Vector to CSV List}
\usage{
code_vector_to_csv_list(vector, add.quotes = TRUE,
  enclose.in.parenthesis = TRUE)
}
\arguments{
\item{vector}{a vector list}

\item{add.quotes}{put quotes around each item in the list}

\item{enclose.in.parenthesis}{enclose the entire csv list in parenthesis}
}
\description{
Returns a comma-seperated list in a string
}
\examples{

code_vector_to_csv_list(c("A","B","C"))

}
\keyword{coder,}
\keyword{sql}
