% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table.R
\name{db_table}
\alias{db_table}
\title{Database Table}
\usage{
db_table(table_name, database = "")
}
\arguments{
\item{database}{the database where the table is located}
}
\description{
State descriptions linked with AICODE
}
\keyword{database,}
\keyword{location}
\keyword{state,}

