% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_combos.R
\name{make_all_combos}
\alias{make_all_combos}
\title{Make All Combinations of a List of Items}
\usage{
make_all_combos(vector, include_these_in_each = NULL)
}
\arguments{
\item{vector}{the list of items}

\item{include_these_in_each}{items that you want to make sure are 
included in each returned itme}
}
\description{
Returns all possible combinations of a list optionally appending 
elements to make sure they are included in each list item
}
\examples{

make_all_combos(c("A","B","C"))
}
