% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_group_by_combos.R
\name{make_all_group_by_combos}
\alias{make_all_group_by_combos}
\title{SQL Group by List}
\usage{
make_all_group_by_combos(group_by_cols)
}
\arguments{
\item{group_by_cols}{The columns that will be included in the group 
by statements}
}
\description{
Turns a character list into a list of group by statements 
that captures every combination of items in the vector
}
