% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_csv_list_from_vector.R
\name{make_csv_list_from_vector}
\alias{make_csv_list_from_vector}
\title{Adapt Vector to CSV List}
\usage{
make_csv_list_from_vector(
  vector,
  add_quotes = TRUE,
  enclose_in_parenthesis = TRUE
)
}
\arguments{
\item{vector}{a vector list}

\item{add_quotes}{put quotes around each item in the list}

\item{enclose_in_parenthesis}{enclose the entire csv list in parenthesis}
}
\description{
Returns a comma-seperated list in a string
}
\examples{

make_csv_list_from_vector(c("A","B","C"))

}
\keyword{coder,}
\keyword{sql}
