% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vector_from_csv_list.R
\name{make_vector_from_csv_list}
\alias{make_vector_from_csv_list}
\title{Adapt CSV List to Vector}
\usage{
make_vector_from_csv_list(csv_list)
}
\arguments{
\item{csv_list}{comma-seperated list in character format}
}
\description{
Returns a vector
}
\keyword{coder,}
\keyword{sql}
