% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_count.R
\name{code_sql_count}
\alias{code_sql_count}
\title{Count Records}
\usage{
code_sql_count(sql, by = NULL)
}
\arguments{
\item{sql}{an object of type {\link{code_sql_sample}}, character in SQL format or a {\link{db_table}} that has SQL which selects records}

\item{by}{count by these groups}
}
\description{
SQL Code to get a distinct column values
}
\details{
Counts the rows from a SELECT statement or table
}
\keyword{database,}
\keyword{location}
\keyword{state,}

