% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_distinct.R
\name{code_sql_distinct}
\alias{code_sql_distinct}
\title{Distinct Records}
\usage{
code_sql_distinct(sql, col)
}
\arguments{
\item{sql}{an object of type {\link{code_sql_sample}}, character in SQL format or a {\link{db_table}} that has SQL which selects records}

\item{col}{the columns that you want to look at}
}
\description{
SQL Code to get a distinct column values
}
\keyword{database,}
\keyword{location}
\keyword{state,}

