% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_sample.R
\name{code_sql_sample}
\alias{code_sql_sample}
\title{Sample}
\usage{
code_sql_sample(table, filter = NULL)
}
\arguments{
\item{table}{indicates the table where the records are located}

\item{filter}{optional filter to limit the records sampled}
}
\description{
SQL Code to get a filtered sample of records
}
\keyword{database,}
\keyword{location}
\keyword{state,}

