% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_select.R
\name{code_sql_select}
\alias{code_sql_select}
\title{Code a SQL SELECT}
\usage{
code_sql_select(table_or_sql, select.cols = c("*"), joins = NULL,
  filters = NULL, filters.logical = "AND", group.by.cols = NULL,
  order.by.cols = NULL, alias = "RECORDS")
}
\arguments{
\item{table_or_sql}{SQL query or table name that you will select records from}

\item{select.cols}{the columns that you would like to select}
}
\description{
Returns a formated sql SELECT query based on parameters. Useful when automating code
}
\keyword{coder,}
\keyword{query}
\keyword{sql,}

