% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_summarize.R
\name{code_sql_summarize}
\alias{code_sql_summarize}
\title{Summarize Records}
\usage{
code_sql_summarize(sql, cols, by = NULL, include_sdp = FALSE)
}
\arguments{
\item{sql}{an object of type {\link{code_sql_sample}}, character in SQL format that has SQL which selects records}

\item{cols}{the columns that you want to analyze}

\item{by}{summarize by these groups}

\item{include_sdp}{include the standard deviation (population). This is optional because some databases (SQLite in particular) do not support this statistic.}
}
\description{
SQL Code to get descriptive statistics
}
\details{
Summarizes the numeric columns from a SELECT statement or table
}
\keyword{stats}

