% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_head.R
\name{sql_head}
\alias{sql_head}
\title{Get Top Records}
\usage{
sql_head(sql)
}
\arguments{
\item{sql}{an object of type sql_sample, character in SQL format or a db_table that has code which selects a sample of records}
}
\description{
SQL Code to get the top five records from a sample
}
\keyword{database,}
\keyword{location}
\keyword{state,}

